requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
tornado==6.4.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
